backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["John Owolabi Idogun <sirneij@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
python-decouple = "^3.6"
django-cors-headers = "^3.13.0"
celery = {extras = ["redis"], version = "^5.2.7"}
psycopg2 = "^2.9.5"
pillow = "^9.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
